change to working elektronn3 branch, update and improve dependency restrictions, remove deprecation warnings

+++ scripts/misc/prepare.py
@@ -67,13 +67,13 @@ if __name__ == '__main__':
 
         # store subgraph of SV-agglomeration
         g_p = "{}/glia/neuron_rag.bz2".format(global_params.config.working_dir)
-        rag_g = nx.read_edgelist(g_p, nodetype=np.uint)
+        rag_g = nx.read_edgelist(g_p, nodetype=np.uint64)
         sv_ids = np.unique(seg)
         rag_sub_g = rag_g.subgraph(sv_ids)
         print('Writing subgraph within {} and {} SVs.'.format(
             bb, rag_sub_g.number_of_nodes()))
         g_p = "{}/glia/glia_rag.bz2".format(global_params.config.working_dir)
-        rag_g = nx.read_edgelist(g_p, nodetype=np.uint)
+        rag_g = nx.read_edgelist(g_p, nodetype=np.uint64)
         rag_sub_g_glia = rag_g.subgraph(sv_ids)
         rag_sub_g = nx.union(rag_sub_g, rag_sub_g_glia)
         os.makedirs(data_dir, exist_ok=True)
